+++ client/src/components/PatientDetails/PatientDetails.css
@@ -4,7 +4,7 @@
   box-shadow: 0 4px 8px rgba(255, 255, 255, 0.1);
   padding: 20px;
   margin-bottom: 20px;
-  width: 25%;
+  width: 100%;
   float: left;
   color: #ffffff;
   min-height: 261px;
