+++ client/src/components/PreviousNotes/PreviousNotes.css
@@ -3,9 +3,10 @@
   padding: 5px;
 }
 .notes-container {
-  width: 85%;
+  width:100%;
+  padding: 3%;
   margin: auto;
-  margin-top: 45px;
+  margin-top: 15px;
   background-color: #47a8bd;
   color: #ffffff;
 }
@@ -15,4 +16,9 @@
   margin-top: 25px;
   background-color: #47a8bd;
   color: #ffffff;
+}
+
+.notes-header, .no-notes {
+  color: #47a8bd;
+  font-size: 35px;
 }
