+++ client/src/components/Room/Room.module.css
@@ -63,4 +63,5 @@
 
   .border {
     border: black;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 13px 27px -5px, rgba(0, 0, 0, 0.3) 0px 8px 16px -8px;
   }
