+++ client/src/components/RoomList/RoomList.module.css
@@ -6,4 +6,5 @@
     flex-direction: column;
     gap: 15px;
     min-height: 80vh;
+    box-shadow: rgba(50, 50, 93, 0.25) 0px 13px 27px -5px, rgba(0, 0, 0, 0.3) 0px 8px 16px -8px;
 }
