--- client/src/pages/Tab2.module.css
@@ -1,5 +1,0 @@
-textarea {
-  background-color: #e1eaec;
-  width: 65%;
-  margin-top: 8px;
-}
