--- client/src/pages/style/style.css
@@ -1,3 +1,0 @@
-.Critical {
-    background-color: red;
-}
